+++ R/utils_checks.R
@@ -1,3 +1,32 @@
+#' Check if a matrix is square (and require the same row and colnames).
+#'
+#' A 'softer' version of isSymmetric that does not require the values to be the same,
+#' only the dimensions and dimnames.
+#'
+#' @keywords internal
+#'
+#' @param x A matrix or data frame to check.
+#'
+#' @returns Logical.
+#'
+isSquare <- function(x) {
+  UseMethod("isSquare")
+}
+
+#' @export
+isSquare.matrix <- function(x) {
+  if (nrow(x) != ncol(x)) return(FALSE)
+  if (is.null(rownames(x)) && !is.null(colnames(x))) return(FALSE)
+  if (!is.null(rownames(x)) && is.null(colnames(x))) return(FALSE)
+  if (!identical(rownames(x), colnames(x))) return(FALSE)
+  return(TRUE)
+}
+
+#' @export
+isSquare.data.frame <- function(x) {
+  isSquare(as.matrix(x))
+}
+
 #' Check that layout and mode are correct
 #'
 #' @keywords internal
@@ -346,4 +375,19 @@ check_annot_names_deprecated <- function(new_params = NULL, old_params = NULL, c
   }
 
   return(params_out)
+}
+
+
+check_xy_deprecated <- function(xy_new, xy_old, old_name) {
+  # Since default is not NULL for these arguments, use the old one if provided
+  if (!is.null(xy_old)) {
+    new_name <- gsub("x", "cols", gsub("y", "rows", old_name))
+    cli::cli_warn("{.var {old_name}} was deprecated in ggcorrheatmap version 0.3.0. Please use {.var {new_name}} instead.")
+
+    xy_out <- xy_old
+  } else {
+    xy_out <- xy_new
+  }
+
+  return(xy_out)
 }