+++ NEWS.md
@@ -1,3 +1,25 @@
+# ggcorrheatmap 0.3.0
+
+## New features
+
+* The new `split_diag` argument allows for triangular cells in the diagonal.
+
+* Square matrices where the row- and column names are the same now work with features that previously required a symmetric matrix (such as triangular layouts).
+
+## Minor changes
+
+* `return_data` now always returns a `layout` column.
+
+* The legends in `ggcorrhm()` no longer assume that Pearson correlation was used when `cor_in` is `TRUE`.
+
+## Fixes
+
+* Annotations and mixed layouts now work with ggplot2 version 4.0.0.
+
+## Deprecated arguments
+
+* `show_names_x/y` and `names_x/y_side` have been deprecated in favour of `show_names_cols/rows` and `names_cols/rows_side` for consistency with other arguments.
+
 # ggcorrheatmap 0.2.0
 
 ## New features